figured out what was going wrong, I was converting an unsigned int to int, whoops

+++ getsys.h
@@ -29,10 +29,10 @@ int GetHostRamInBytes(void)
 }
 
 
-int GetDeviceRamInBytes(void){
+unsigned int GetDeviceRamInBytes(void){
     size_t free, total;
     cudaMemGetInfo(&free, &total);
-    return (int) free;
+    return free;
 }
 
 #endif
